refactor: update older folder instead of creating a new one
