+++ packages/onboarding-ui/src/pages/CreateCloudWorkspacePage/TitleCreateCloudPage.tsx
@@ -4,9 +4,9 @@ import { Trans } from 'react-i18next';
 import { TitleHighlight } from '../../common/FormPageLayout.styles';
 
 const TitleCreateCloudPage = (): ReactElement => (
-  <Trans i18nKey='page.cloudDescription.title'>
-    Let's launch your workspace and
-    <TitleHighlight>14-day trial Launch</TitleHighlight>
+  <Trans i18nKey='page.createCloudWorkspace.title'>
+    Launch new workspace and
+    <TitleHighlight>30-day trial</TitleHighlight>
   </Trans>
 );
