--- packages/onboarding-ui/src/pages/CreateCloudWorkspaceSimplifiedPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CreateCloudWorkspaceSimplifiedPage';
